docker/ui/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pyyaml>=6.0
openai>=1.12.0
chromadb>=0.4.22
sentence-transformers>=2.6.0
